# openapi.yaml
openapi: 3.0.3
info:
  title: Bookings Engine API
  description: |
    Generic multi-tenant booking API for service-based businesses.
    Supports gyms, restaurants, salons, nail bars, and spas through
    configurable business rules rather than business-specific code paths.
  version: 1.0.0
  contact:
    name: Veltrion
servers:
  - url: https://api.bookings.example.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /resource-types:
    get:
      tags: [Resource Types]
      summary: List resource types
      operationId: listResourceTypes
      responses:
        "200":
          description: List of resource types
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResourceType"
    post:
      tags: [Resource Types]
      summary: Create a resource type
      operationId: createResourceType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceTypeCreate"
      responses:
        "201":
          description: Created resource type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceType"
  /resource-types/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceTypeId"
    patch:
      tags: [Resource Types]
      summary: Update a resource type
      operationId: updateResourceType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceTypeUpdate"
      responses:
        "200":
          description: Updated resource type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceType"
    delete:
      tags: [Resource Types]
      summary: Delete a resource type
      operationId: deleteResourceType
      responses:
        "204":
          description: Deleted

  # ─── Resources ─────────────────────────────────────────────────────────────────
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      operationId: listCustomers
      parameters:
        - name: search
          in: query
          description: Search by name, email, or phone
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Paginated list of customers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Customer"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Customers]
      summary: Create a customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerCreate"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
    get:
      tags: [Customers]
      summary: Get a customer
      operationId: getCustomer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      tags: [Customers]
      summary: Update a customer
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ─── Bookings ──────────────────────────────────────────────────────────────────
  /bookings:
    get:
      tags: [Bookings]
      summary: List bookings
      operationId: listBookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled, no_show]
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: resource_id
          in: query
          schema:
            type: string
            format: uuid
        - name: customer_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Paginated list of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Bookings]
      summary: Create a booking
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCreate"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Time slot no longer available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookings/{id}:
    parameters:
      - $ref: "#/components/parameters/BookingId"
    get:
      tags: [Bookings]
      summary: Get a booking
      operationId: getBooking
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      tags: [Bookings]
      summary: Update a booking
      operationId: updateBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Conflict when rescheduling
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookings/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/BookingId"
    post:
      tags: [Bookings]
      summary: Cancel a booking
      operationId: cancelBooking
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Cancelled booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "403":
          description: Cancellation window expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookings/{id}/complete:
    parameters:
      - $ref: "#/components/parameters/BookingId"
    post:
      tags: [Bookings]
      summary: Mark booking as completed
      operationId: completeBooking
      responses:
        "200":
          description: Completed booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/no-show:
    parameters:
      - $ref: "#/components/parameters/BookingId"
    post:
      tags: [Bookings]
      summary: Mark booking as no-show
      operationId: noShowBooking
      responses:
        "200":
          description: No-show booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceTypeId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CustomerId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BookingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
  schemas:
...
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, NOT_FOUND, CONFLICT, FORBIDDEN, UNAUTHORIZED]
            message:
              type: string
            details:
              type: object

    # ─── Tenant ────────────────────────────────────────────────────────────────
...
    ResourceType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
    ResourceTypeCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    ResourceTypeUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    # ─── Resources ─────────────────────────────────────────────────────────────
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        notes:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CustomerCreate:
      type: object
      required: [first_name, last_name]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        notes:
          type: string
        metadata:
          type: object
    CustomerUpdate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        notes:
          type: string
        metadata:
          type: object

    # ─── Bookings ──────────────────────────────────────────────────────────────
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no_show]
        service:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        resource:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        customer:
          type: object
          properties:
            id:
              type: string
              format: uuid
            first_name:
              type: string
            last_name:
              type: string
        start_time:
          type: string
          format: date-time
          description: UTC
        start_time_local:
          type: string
          description: Tenant local time
        end_time:
          type: string
          format: date-time
          description: UTC
        end_time_local:
          type: string
          description: Tenant local time
        party_size:
          type: integer
        notes:
          type: string
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        cancellation_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BookingCreate:
      type: object
      required: [service_id, customer_id, start_time]
      properties:
        service_id:
          type: string
          format: uuid
        resource_id:
          type: string
          format: uuid
          nullable: true
          description: "Omit or null for 'any available' assignment"
        customer_id:
          type: string
          format: uuid
        start_time:
          type: string
          description: "ISO 8601 datetime in tenant's local timezone"
        party_size:
          type: integer
          minimum: 1
          default: 1
        notes:
          type: string
    BookingUpdate:
      type: object
      properties:
        start_time:
          type: string
          description: "Reschedule: new start time in tenant's local timezone"
        resource_id:
          type: string
          format: uuid
          description: "Reassign to different resource"
        notes:
          type: string